fix(search): moved yearrangeslider to seperate file
